refactor: eulexia fab component split into fab and action components
