+++ src/index.js
@@ -1,3 +1,15 @@
-import EulexiaFab from './EulexiaFab/index.jsx'
+import EulexiaFab from './EulexiaFab/EulexiaFab.jsx'
+import ColorChangeAction from './EulexiaAction/ColorChangeAction/ColorChangeAction.jsx'
+import FontFamilyAction from './EulexiaAction/FontFamilyAction/FontFamilyAction.jsx'
+import FontSizeAction from './EulexiaAction/FontSizeAction/FontSizeAction.jsx'
+import RulerAction from './EulexiaAction/RulerAction/RulerAction.jsx'
+import TextToSpeechAction from './EulexiaAction/TextToSpeechAction/TextToSpeechAction.jsx'
 
-export { EulexiaFab }
+export {
+  EulexiaFab,
+  ColorChangeAction,
+  FontFamilyAction,
+  FontSizeAction,
+  RulerAction,
+  TextToSpeechAction
+}
